source/src/main/java/br/com/uol/pagseguro/api/common/domain/converter/BankV3MapConverter.java
/*
 * 2007-2016 [PagSeguro Internet Ltda.]
 * 
 * NOTICE OF LICENSE
 * 
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 * http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * 
 * Copyright: 2007-2016 PagSeguro Internet Ltda.
 * Licence: http://www.apache.org/licenses/LICENSE-2.0
 */

package br.com.uol.pagseguro.api.common.domain.converter;

import br.com.uol.pagseguro.api.common.domain.Bank;
import br.com.uol.pagseguro.api.utils.AbstractMapConverter;
import br.com.uol.pagseguro.api.utils.RequestMap;

/**
 * Converter for bank. Used in version 3.
 * Used to convert attributes of bank in request map.
 *
 * @author PagSeguro Internet Ltda.
 */
public class BankV3MapConverter extends AbstractMapConverter<Bank> {

  /**
   * Convert attributes of bank in request map
   *
   * @param requestMap Request Map used to pass params to api
   * @param bank       The interface of bank
   * @see RequestMap
   * @see Bank
   * @see AbstractMapConverter#convert(Object)
   */
  @Override
  protected void convert(RequestMap requestMap, Bank bank) {
    requestMap.putString("bank.name", bank.getBankName().getStringName());
  }
}
